读取资源逻辑大改

+++ sf/include/TArray.h
@@ -1,19 +1,19 @@
 #ifndef TARRAY_H
 #define TARRAY_H
 #pragma execution_character_set("utf-8")
-#include <TInit.h>
+#include <init.h>
 
 template<class MyBMPixel>
 class TArray
 {
 public:
 	TArray();
-	TArray(size_t size);
+	TArray(unsigned int size);
 	~TArray();
-	MyBMPixel& operator[](size_t id);
+	MyBMPixel& operator[](unsigned int id);
 
 protected:
-	size_t m_size;
+	unsigned int m_size;
 	MyBMPixel* m_pArrHead;
 
 private:
@@ -26,7 +26,7 @@ TArray<MyBMPixel>::TArray()
 }
 
 template<class MyBMPixel>
-TArray<MyBMPixel>::TArray(size_t size)
+TArray<MyBMPixel>::TArray(unsigned int size)
 {
 	m_size = size;
 	m_pArrHead = new MyBMPixel[size];
@@ -39,7 +39,7 @@ TArray<MyBMPixel>::~TArray()
 }
 
 template<class MyBMPixel>
-MyBMPixel& TArray<MyBMPixel>::operator[](size_t id)
+MyBMPixel& TArray<MyBMPixel>::operator[](unsigned int id)
 {
 	return m_pArrHead[id];
 }
